pyproject.toml
[project]
name = "smarthome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "awsiotsdk>=1.28.0",
    "boto3>=1.42.39",
    "fastmcp>=2.14.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tapo>=0.8.7",
]

[dependency-groups]
dev = [
    "moto[dynamodb]>=5.1.20",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
